httpstats: promote context tags to stats package
